Bigger changes: Added the Coastline package. Coastline uses a special file format that we call ".dat". CustomFilter now changes all Node IDs in the XML to 1 and up, in increasing order. This is to avoid common node IDs between the Coastline data and the OSM data.
GeoConvert is a Class used to transform Lat/Lon to UTM32, to match Kraks coordinate system. I did not write this class, but I stripped the class from all unnecessary methods and altered the toUTM class to be forced to a specific UTM to avoid problems when the lat/lon are outside of one UTM zone.

Dataloader now loads Coastline after loading the Krak or OSM data.

Minor changes:
Added a line to avoid null pointer references in QuadTree. (if fn != null && tn != null)
Added setX and setY methods to the MapNode interface.